pyproject.toml
[tool.poetry]
name = "Horde Workspace"
version = "0.1.0"
description = "Utilities for generating images via the AI Horde."
authors = ["Luke100000 <jonas.peche@aon.at>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.10"
horde-sdk = "^0.14.1"
numpy = "^2.1.0"
pillow = "^10.4.0"
aiohttp = "^3.10.5"
pydantic-yaml = "^1.3.0"
scikit-learn = "^1.5.1"
scikit-image = "^0.24.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
